+++ Runtime/System/Splines/NCurve.cs
@@ -5,26 +5,38 @@ using UnityEngine;
 namespace Nazio_LT.Tools.Core
 {
     [System.Serializable]
-    public class NSplineHandle
+    public class NHandle
     {
         public Vector3 point;
-        public Vector3 helper1;
-        public Vector3 helper2;
+        public Vector3 forwardHelper;
+        public Vector3 backHelper;
     }
 
     [System.Serializable]
     public sealed class NCurve
     {
-        private enum CurveType { Linear = 0, Bezier = 1}
-        [SerializeField] private List<NSplineHandle> handles = new List<NSplineHandle>();
+        private enum CurveType { Linear = 0, Bezier = 1 }
+        [SerializeField] private CurveType type;
+        [SerializeField] private List<NHandle> handles = new List<NHandle>();
 
         public Vector3 ComputePoint(float _t)
         {
-            if(_t == 0) _t = 0.001f;
+            if (_t == 0) _t = 0.001f;
             float _computedT = _t / Factor;
             int _curveID = (int)_computedT;
 
-            return Vector3.Lerp(handles[_curveID].point, handles[GetNextID(_curveID)].point, _computedT - _curveID);
+            return GetPointFunc()(handles[_curveID], handles[GetNextID(_curveID)], _computedT - _curveID);
+        }
+
+        private System.Func<NHandle, NHandle, float, Vector3> GetPointFunc()
+        {
+            switch(type)
+            {
+                case CurveType.Bezier:
+                    return (_h1, _h2, _t) => NMath.BezierLerp(_h1, _h2, _t);
+            }
+
+            return (_h1, _h2, _t) => Vector3.Lerp(_h1.point, _h2.point, _t);
         }
 
         private int GetNextID(int _value)